Update walletConnect configuration
