+++ src/core/providers/types/providerFactory.types.ts
@@ -1,5 +1,4 @@
 import type { IDAppProviderBase } from '@multiversx/sdk-dapp-utils';
-import { WalletConnectV2ProviderOptionsType } from '@multiversx/sdk-wallet-connect-provider/out';
 
 // @ts-ignore
 export interface IProvider<T extends ProviderTypeEnum = ProviderTypeEnum>
@@ -42,13 +41,6 @@ export interface IProviderConfig {
     address: string;
   };
   UI?: IProviderConfigUI;
-  walletConnect?: {
-    walletConnectV2ProjectId: string;
-    walletConnectV2RelayAddress?: string;
-    walletConnectV2Options?: WalletConnectV2ProviderOptionsType;
-    customRequestMethods?: Array<string>;
-    logoutRoute?: string;
-  };
 }
 
 export enum ProviderTypeEnum {
@@ -67,7 +59,6 @@ export interface IProviderFactory<
   T extends ProviderTypeEnum = ProviderTypeEnum
 > {
   type: T[keyof T];
-  config: IProviderConfig;
 }
 
 export interface ICustomProvider<
