+++ src/store/actions/config/configActions.ts
@@ -1,4 +1,4 @@
-import { IProviderConfig } from 'core/providers/types/providerFactory.types';
+import { WalletConnectConfig } from 'core/providers/helpers/walletConnect/walletConnect.types';
 import { NativeAuthConfigType } from 'services/nativeAuth/nativeAuth.types';
 import { getStore } from 'store/store';
 
@@ -7,9 +7,7 @@ export const setNativeAuthConfig = (config: NativeAuthConfigType) =>
     state.nativeAuthConfig = config;
   });
 
-export const setWalletConnectConfig = (
-  config: IProviderConfig['walletConnect']
-) =>
+export const setWalletConnectConfig = (config: WalletConnectConfig) =>
   getStore().setState(({ config: state }) => {
     state.walletConnectConfig = config;
   });
