+++ src/store/slices/config/config.types.ts
@@ -1,7 +1,7 @@
-import { IProviderConfig } from 'core/providers/types/providerFactory.types';
+import { WalletConnectConfig } from 'core/providers/helpers/walletConnect/walletConnect.types';
 import { NativeAuthConfigType } from 'services/nativeAuth/nativeAuth.types';
 
 export interface ConfigSliceType {
   nativeAuthConfig: NativeAuthConfigType | null;
-  walletConnectConfig: IProviderConfig['walletConnect'] | null;
+  walletConnectConfig: WalletConnectConfig | null;
 }
